fix: cmd/server Упаковать конфигурацию в configure(default)
по аналогии с агентом,
 собрать из файлов flag.go и env.go
